Do not explicity specify max_parallel to addprocs
This has more to do with the SSH server's MaxStarups option than
anything else, but max_parallel doesn't seem to affect the connections
through an SSH gateway so I think it might only affect the connections
to the end nodes themselves (i.e. when nprocs>1), but not any
intervening gateways.